Added Gauss-Legendre Integration Functionality

+++ dependencies/math-extended/src/math_extended.cpp
@@ -1,6 +1,6 @@
 #include "math_extended.h"
 
-int factorial(int n)
+long long int factorial(int n)
 {
     return (n == 1 || n == 0) ? 1 : factorial(n - 1) * n;
 }
