+++ dependencies/math-extended/src/math_extended.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int factorial(int n);
+long long int factorial(int n);
